publisher and publisher message moved to own directory
